+++ include/vendors/AirSpy.h
@@ -35,15 +35,15 @@ namespace PortSDR
         int Stop() override;
         int SetCenterFrequency(uint32_t freq, int stream) override;
         int SetSampleRate(uint32_t sampleRate) override;
-        int SetGain(int gain) override;
+        int SetGain(double gain) override;
         int SetSampleFormat(SampleFormat format) override;
 
-        int SetGain(int gain, std::string_view name) override;
+        int SetGain(double gain, std::string_view name) override;
         int SetGainModes(std::string_view name) override;
 
-        int SetLnaGain(int gain);
-        int SetMixGain(int gain);
-        int SetIfGain(int gain);
+        int SetLnaGain(double gain);
+        int SetMixGain(double gain);
+        int SetIfGain(double gain);
 
         [[nodiscard]] std::vector<uint32_t> GetSampleRates() const override;
         [[nodiscard]] std::vector<SampleFormat> GetSampleFormats() const override;
